# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Eventless API",
    "version": "1.0"
  },
  "paths": {
    "/api/events/{id}/attendee/{email}/qrcode": {
      "get": {
        "summary": "Generates a PDF with event details and a QR code for a registered attendee",
        "description": "This endpoint generates a PDF containing event details and a QR code for an attendee who is registered for the event. The QR code is created using a hash that includes the event ID, attendee email, and attendee ID. It requires a valid API key for authorization.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the event",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "email",
            "required": true,
            "description": "The email address of the attendee",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "api-key",
            "required": true,
            "description": "API key for authorization",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully generated PDF with event details and QR code",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bad request due to event or attendee issues",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "NOT_FOUND"
                    },
                    "message": {
                      "type": "string",
                      "example": "Event with id {id} not found or You are not registered for this event"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string",
                      "example": "Your API Key is invalid."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "INTERNAL_SERVER_ERROR"
                    },
                    "message": {
                      "type": "string",
                      "example": "please wait a moment and try again"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{id}/attendee": {
      "post": {
        "summary": "Registers an attendee for an event with answers to required questions",
        "description": "This endpoint registers an attendee for a specific event. It checks if the event exists, if registration is still open, if the attendee is already registered, and if all required questions have been answered. It requires a valid API key for authorization.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the event to register for",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "api-key",
            "required": true,
            "description": "API key for authorization",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "Registration",
            "description": "The registration details including email and answers to required questions",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "description": "The email address of the attendee",
                      "example": "attendee@example.com"
                    },
                    "answers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "questionId": {
                            "type": "string",
                            "description": "The ID of the question",
                            "example": "q1"
                          },
                          "answerText": {
                            "type": "string",
                            "description": "The answer text for the question",
                            "example": "Yes"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully registered the attendee",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "The ID of the newly created attendee",
                      "example": "attendee123"
                    },
                    "email": {
                      "type": "string",
                      "description": "The email address of the registered attendee",
                      "example": "attendee@example.com"
                    },
                    "answers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "questionId": {
                            "type": "string",
                            "description": "The ID of the question",
                            "example": "q1"
                          },
                          "answerText": {
                            "type": "string",
                            "description": "The answer text for the question",
                            "example": "Yes"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request due to issues such as event not found, registration deadline exceeded, already registered, or missing required answers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "NOT_FOUND"
                    },
                    "message": {
                      "type": "string",
                      "example": "Event with id {id} not found or Registration is not possible anymore. deadline ended on {registrationEndDate}"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string",
                      "example": "Your API Key is invalid."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "INTERNAL_SERVER_ERROR"
                    },
                    "message": {
                      "type": "string",
                      "example": "please wait a moment and try again"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}